evalGemini/pyproject.toml
[project]
name = "browsernerd-eval-gemini"
version = "0.1.0"
description = "Evaluation harness for BrowserNERD MCP vs baselines (Gemini version)"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0.0",
    "mcp>=1.1.0",
    "httpx>=0.27.0",
    "playwright>=1.48.0",
    "pydantic>=2.9.0",
    "scipy>=1.14.0",
    "python-dotenv>=1.2.1",
]

[project.scripts]
browsernerd-eval-gemini = "cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = [
    "*.py",
    "runners/*.py",
]
